generate random numbers as floats
as they get converted from double before use anyway

+++ OpenRA.Mods.RA/AutoTarget.cs
@@ -107,7 +107,7 @@ namespace OpenRA.Mods.RA
 		{
 			var info = self.Info.Traits.Get<AttackBaseInfo>();
 			nextScanTime = (int)(25 * (info.ScanTimeAverage +
-				(self.World.SharedRandom.NextDouble() * 2 - 1) * info.ScanTimeSpread));
+				(self.World.SharedRandom.NextFloat() * 2 - 1) * info.ScanTimeSpread));
 
 			var inRange = self.World.FindUnitsInCircle(self.CenterLocation, (int)(Game.CellSize * range));
